# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Development server"
        }
    ],
    "paths": {
        "/books": {
            "get": {
                "tags": [
                    "ApiBook"
                ],
                "summary": "ApiBook show all",
                "operationId": "eda618c428b3bfd19b43b5c87bd893f5",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "ApiBook"
                ],
                "summary": "ApiBook create",
                "operationId": "71b0ad0bdf7a9cbbbf0550fd4d1f8637",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "author",
                                    "pages"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "War and Peace"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "L. N. Tolstoy"
                                    },
                                    "pages": {
                                        "type": "integer",
                                        "example": "1300"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "ApiBook"
                ],
                "summary": "ApiBook show",
                "operationId": "da881e6bd02022383737556c5f6be9e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "ApiBook"
                ],
                "summary": "ApiBook update",
                "operationId": "db7d66b491dfbf04fc34692b1637b38f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated book object",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ApiBook"
                ],
                "summary": "ApiBook destroy",
                "operationId": "d11c0821154aec6bab7ff7d36b23d1fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Example"
                ],
                "summary": "Get users",
                "operationId": "9dd33c03c21315e4d87351e76fd287bc",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Example",
            "description": "Example API"
        },
        {
            "name": "ApiBook",
            "description": "Books API"
        }
    ]
}